requirements_render.txt
Django>=5.0,<5.2
django-cors-headers>=4.3.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
drf-yasg>=1.21.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
dj-database-url>=2.1.0
whitenoise>=6.6.0
gunicorn>=21.0.0
requests>=2.31.0
pillow>=10.0.0
cryptography>=41.0.0
pyotp>=2.9.0
urllib3>=2.0.0
certifi>=2023.7.22
PyJWT>=2.8.0
django-filter>=23.0
google-genai>=1.0.0
duckduckgo-search>=4.0.0
rich>=13.0.0